% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd_AddWeights.R
\name{ggd_AddWeights}
\alias{ggd_AddWeights}
\title{Title}
\usage{
ggd_AddWeights(
  data,
  reference,
  by.x = c("GEMEENTECODE", "OnderwijsSoort", "GESLACHT", "MBOKK331"),
  by.y = c("Gemeente", "OnderwijsSoort", "Geslacht", "Leerjaar"),
  by.y.filters = list(c(153, 141, 164, 1742, 163, 173, 189, 158, 183, 1735, 1700, 1774,
    168, 147), NA, c(1, 2), c(2, 4)),
  respondent.id = "Respondentnummer",
  weight.limits = c(0.2, 5),
  force.cast.to.character = "Gemeente",
  append = T,
  do.rescale = T
)
}
\arguments{
\item{data}{yay}

\item{reference}{explain}

\item{by.x}{this is x}

\item{by.y}{this is y}

\item{by.y.filters}{etc}

\item{respondent.id}{etc}

\item{weight.limits}{etc}

\item{force.cast.to.character}{etc laksfh}

\item{append}{etc}

\item{do.rescale}{etc}
}
\value{
describe output
}
\description{
Title
}
\examples{
write example of inputs and output of function
}
