% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd_generateMetaData.R
\name{ggd_generateMetadata}
\alias{ggd_generateMetadata}
\title{Title title of function}
\usage{
ggd_generateMetadata(
  data.path,
  id.prefix,
  metadata.excel.path = "",
  target.path = ""
)
}
\arguments{
\item{data.path}{explain this parameter}

\item{id.prefix}{explain this parameter}

\item{metadata.excel.path}{explain this parameter}

\item{target.path}{explain this parameter}
}
\value{
function returns x
}
\description{
Title title of function
}
\examples{
write one or more examples here that describe the function
}
