% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd_getClusterVariables.R
\name{ggd_getClusterVariables}
\alias{ggd_getClusterVariables}
\title{Title title of function}
\usage{
ggd_getClusterVariables(
  data,
  outcome,
  method = "all",
  desired.n = 10,
  desired.n.relax = F,
  selected.variables = NA,
  glm.formula = "y ~.",
  glm.p.value.threshold = 0.05,
  rf.num.trees = 1000,
  rf.num.try = 1,
  rf.min.node.size = 0.1,
  swr.alpha.to.enter = 0.15,
  swr.alpha.to.remove = 0.15,
  swr.n.iter = 100
)
}
\arguments{
\item{data}{explain parameter}

\item{outcome}{explain parameter}

\item{method}{explain parameter}

\item{desired.n}{explain parameter}

\item{desired.n.relax}{explain parameter}

\item{selected.variables}{explain parameter}

\item{glm.formula}{explain parameter}

\item{glm.p.value.threshold}{explain parameter}

\item{rf.num.trees}{explain parameter}

\item{rf.num.try}{explain parameter}

\item{rf.min.node.size}{explain parameter}

\item{swr.alpha.to.enter}{explain parameter}

\item{swr.alpha.to.remove}{explain parameter}

\item{swr.n.iter}{explain parameter}
}
\value{
function returns x
}
\description{
Title title of function
}
\examples{
write one or more examples here that describe the ggd_AddWeights function
}
