% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd_glm.R
\name{ggd_glm}
\alias{ggd_glm}
\title{Title title of function}
\usage{
ggd_glm(y, x, glm.formula, glm.p.value.threshold)
}
\arguments{
\item{y}{explain parameter}

\item{x}{explain parameter}

\item{glm.formula}{explain parameter}

\item{glm.p.value.threshold}{explain parameter}
}
\value{
function returns x
}
\description{
Title title of function
}
\examples{
write one or more examples here that describe the ggd_AddWeights function
}
