% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd_hierarchicalCluster.R
\name{ggd_hierarchicalCluster}
\alias{ggd_hierarchicalCluster}
\title{Title title of function}
\usage{
ggd_hierarchicalCluster(
  data,
  n.clusters,
  distance.metric = "euclidean",
  method.in = "average",
  append.clusters = T
)
}
\arguments{
\item{data}{explain parameter}

\item{n.clusters}{explain parameter}

\item{distance.metric}{explain parameter}

\item{method.in}{explain parameter}

\item{append.clusters}{explain parameter}
}
\value{
function returns x
}
\description{
Title title of function
}
\examples{
write one or more examples here that describe the ggd_AddWeights function
}
