% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd_loadData.R
\name{ggd_loadData}
\alias{ggd_loadData}
\title{Title title of function}
\usage{
ggd_loadData(
  data.path,
  metadata.path,
  select.columns = "*",
  select.rows,
  set.spss.missings.to.na = T
)
}
\arguments{
\item{data.path}{explain this parameter}

\item{metadata.path}{explain this parameter}

\item{select.columns}{explain this parameter}

\item{select.rows}{explain this parameter}

\item{set.spss.missings.to.na}{explain this parameter}
}
\value{
function returns x
}
\description{
Title title of function
}
\examples{
write one or more examples here that describe the function
}
