% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd_makeCategoricDataOverview.R,
%   R/ggd_make_categoric_data_overview_copy.R
\name{ggd_make_categoric_data_overview}
\alias{ggd_make_categoric_data_overview}
\title{Title title of function}
\usage{
ggd_make_categoric_data_overview(
  dataset,
  name_dataset = "",
  output_frequency_table = FALSE,
  output_scale_factor_table = FALSE,
  output_score_per_row = FALSE
)

ggd_make_categoric_data_overview(
  dataset,
  name_dataset = "",
  output_frequency_table = FALSE,
  output_scale_factor_table = FALSE,
  output_score_per_row = FALSE
)
}
\arguments{
\item{dataset}{explain parameter}

\item{name_dataset}{explain parameter}

\item{output_frequency_table}{explain parameter}

\item{output_scale_factor_table}{explain parameter}

\item{output_score_per_row}{explain parameter}
}
\value{
function returns x

function returns x
}
\description{
Title title of function

Title title of function
}
\examples{
write one or more examples here that describe the ggd_AddWeights function
write one or more examples here that describe the ggd_AddWeights function
}
