% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd_makeFrequencyTable.R,
%   R/ggd_make_frequency_table_copy.R
\name{ggd_make_frequency_table}
\alias{ggd_make_frequency_table}
\title{Title title of function}
\usage{
ggd_make_frequency_table(
  data,
  remove_zeroes_row = FALSE,
  remove_zeroes_column = FALSE
)

ggd_make_frequency_table(
  data,
  remove_zeroes_row = FALSE,
  remove_zeroes_column = FALSE
)
}
\arguments{
\item{data}{explain parameter}

\item{remove_zeroes_row}{explain parameter}

\item{remove_zeroes_column}{explain parameter}
}
\value{
function returns x

function returns x
}
\description{
Title title of function

Title title of function
}
\examples{
write one or more examples here that describe the ggd_AddWeights function
write one or more examples here that describe the ggd_AddWeights function
}
