% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd_randomInpute.R
\name{ggd_randomImpute}
\alias{ggd_randomImpute}
\title{Title title of function}
\usage{
ggd_randomImpute(data, specified.columns)
}
\arguments{
\item{data}{explain parameter}

\item{specified.columns}{explain parameter}
}
\value{
function returns x
}
\description{
Title title of function
}
\examples{
write one or more examples here that describe the ggd_AddWeights function
}
