% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd_recodeValues.R
\name{ggd_recodeValues}
\alias{ggd_recodeValues}
\title{Title recode values in data}
\usage{
ggd_recodeValues(to_be_recoded, old_coding, new_coding)
}
\arguments{
\item{to_be_recoded}{the dataset that needs to be recoded}

\item{old_coding}{vector of values with the old coding}

\item{new_coding}{vector of values with the new coding}
}
\value{
the function returns the data with the recoded values
}
\description{
Title recode values in data
}
\examples{
nog te doen
}
