% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd_resample.R
\name{ggd_resample}
\alias{ggd_resample}
\title{Title title of function}
\usage{
ggd_resample(data, weights = NA, desired.size = 1)
}
\arguments{
\item{data}{explain parameter}

\item{weights}{explain parameter}

\item{desired.size}{explain parameter}
}
\value{
function returns x
}
\description{
Title title of function
}
\examples{
write one or more examples here that describe the ggd_AddWeights function
}
