% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd_svyStepwiseRegression.R
\name{ggd_SvyStepwiseRegression}
\alias{ggd_SvyStepwiseRegression}
\title{Title title of function}
\usage{
ggd_SvyStepwiseRegression(
  data,
  outcome,
  predictor.subset.xls = NA,
  alpha.to.enter = 0.15,
  alpha.to.remove = 0.15,
  n.iter = 100
)
}
\arguments{
\item{data}{explain parameter}

\item{outcome}{explain parameter}

\item{predictor.subset.xls}{explain parameter}

\item{alpha.to.enter}{explain parameter}

\item{alpha.to.remove}{explain parameter}

\item{n.iter}{explain parameter}
}
\value{
function returns x
}
\description{
Title title of function
}
\examples{
write one or more examples here that describe the ggd_AddWeights function
}
